Modal refactoring
