+++ src/App.js
@@ -1,12 +1,9 @@
-//import logo from './logo.svg';
 import './App.css';
 import { useState } from 'react';
 import {BrowserRouter, Route, Routes} from "react-router-dom";
+
 import Name from "./login/signUp/Name";
-import Main from "./login/MainLoginPage";
-//import Popup from "./login/StaticExample";
-import Modal from './commons/components/Modal.js'
-import Button from 'react-bootstrap/Button';
+import Main from "./MainTest";
 import ToLoginModal from "./login/ToLoginModal.js";
 
 
@@ -24,9 +21,9 @@ function App() {
       <BrowserRouter>
           <Routes>
               <Route path={""} element={<Main />}></Route>
-              <Route path={"/user/Name"} element={<Name />}></Route>
-              <Route path={"/user/Main"} element={<Main />}></Route>
-              <Route path={"/user/ToLoginModal"} element={<ToLoginModal />}></Route>
+              <Route path={"/Name"} element={<Name />}></Route>
+              <Route path={"/Main"} element={<Main />}></Route>
+              <Route path={"/ToLoginModal"} element={<ToLoginModal />}></Route>
           </Routes>
           </BrowserRouter>
