+++ src/commons/components/Modal.js
@@ -1,18 +1,18 @@
 //import styles from '../../assets/css/ModalBasic.module.css';
-//
-//function ModalBasic({ setModalOpen, id, title, content, writer }: PropsType) {
+
+// function ModalBasic() {
 //    // 모달 끄기
-//    const closeModal = () => {
-//        setModalOpen(false);
-//    };
+//    // const closeModal = () => {
+//    //     setModalOpen(false);
+//    // };
 //
 //    return (
 //        <div className={styles.container}>
-//            <button className={styles.close} onClick={closeModal}>
+//            <button className={styles.close} onClick={}>
 //                X
 //            </button>
 //            <p>모달창입니다.</p>
 //        </div>
 //    );
-//}
-//export default ModalBasic;
+// }
+// export default ModalBasic;
